Close #1: Fix .(s)doc dynamically wrt. scroll position
Obviously, it is nice to automatically adjust the classes of the menu
depending on the current scroll position, so we implement it. The
algorithm is very simplistic, and there's room for improvement, but it
seems to be sufficient for now.

To ease the detection of the relevant menu, we add the CSS class
onepage_menu with a hack in Onepage_Li; that might also require
improvement.