Close #11: Make scroll duration dependent on scroll amount
We reinterpret the scroll→duration as *maximum* duration of the
scrolling. The maximum duration is only applied when scrolling the
whole document (top to bottom or vice versa). If only a fraction of the
document has to be scrolled, the scrolling endures only this fraction
of the maximum duration.

We also appropriately increase the default value.